JumpGame #5 スクリプトの整理をしました。GetMicInputが今回新しく作成したものです。 現在、音声の取得とコンソールに音声の音量を表示することが可能です。
