--- TestProject/Assets/Scripts/MicInput.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a373bf9a62a617348ab13a1b967bf413
-timeCreated: 1443161974
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
